gear machineを作れるようにする

+++ moorestech_server/Assets/Scripts/Game.Block/Blocks/Machine/VanillaElectricMachineComponent.cs
@@ -26,7 +26,7 @@ namespace Game.Block.Blocks.Machine
         
         #region IBlockElectric implementation
         
-        public int RequestEnergy => _vanillaMachineProcessorComponent.RequestPower;
+        public float RequestEnergy => _vanillaMachineProcessorComponent.RequestPower;
         
         public void SupplyEnergy(int power)
         {
